// angular/src/app/models/patient.model.ts
export class Patient {
	FirstName: string;
	LastName: string;
	FullName: string;
	FullName20: string;
	MRN: string;
	DateOfBirth: string;
	Gender: string;
	AccountNumber: string;
	Address: string;
	DriversLicense: string;
	MPIID: string;
	NationalHealthID: string;
	PrimaryId: string;
	PrimaryIdFacility: string;
	SSN: string;
	SecondaryId: string;


	constructor() {
		this.FirstName = '';
		this.LastName = '';
		this.FullName = '';
		this.FullName20 = '';
		this.MRN = '';
		this.DateOfBirth = '';
		this.Gender = '';
		this.AccountNumber = '';
		this.Address = '';
		this.DriversLicense = '';
		this.MPIID = '';
		this.NationalHealthID = '';
		this.PrimaryId = '';
		this.PrimaryIdFacility = '';
		this.SSN = '';
		this.SecondaryId = '';
	}
}